+++ ocaml-lsp-server/src/ocaml_lsp_server.ml
@@ -96,6 +96,7 @@ let initialize_info (client_capabilities : ClientCapabilities.t) : InitializeRes
               ; Req_type_enclosing.capability
               ; Req_get_documentation.capability
               ; Req_construct.capability
+              ; Req_type_search.capability
               ] )
         ]
     in
@@ -526,6 +527,7 @@ let on_request
        ; Req_type_enclosing.meth, Req_type_enclosing.on_request
        ; Req_get_documentation.meth, Req_get_documentation.on_request
        ; Req_wrapping_ast_node.meth, Req_wrapping_ast_node.on_request
+       ; Req_type_search.meth, Req_type_search.on_request
        ; Req_construct.meth, Req_construct.on_request
        ; ( Semantic_highlighting.Debug.meth_request_full
          , Semantic_highlighting.Debug.on_request_full )